Move weight extraction tests into their own class
