+++ core/src/test/java/com/graphhopper/routing/util/DataFlagEncoderTest.java
@@ -1,9 +1,34 @@
 package com.graphhopper.routing.util;
 
+import static com.graphhopper.util.GHUtility.updateDistancesFor;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
+
+import java.util.Arrays;
+import java.util.Collections;
+
+import org.junit.Test;
+
 import com.graphhopper.reader.ReaderWay;
-import com.graphhopper.routing.profiles.*;
-import com.graphhopper.routing.util.parsers.*;
-import com.graphhopper.routing.util.parsers.helpers.OSMWeightExtractor;
+import com.graphhopper.routing.profiles.BooleanEncodedValue;
+import com.graphhopper.routing.profiles.Country;
+import com.graphhopper.routing.profiles.DecimalEncodedValue;
+import com.graphhopper.routing.profiles.EnumEncodedValue;
+import com.graphhopper.routing.profiles.IntEncodedValue;
+import com.graphhopper.routing.profiles.MaxSpeed;
+import com.graphhopper.routing.profiles.RoadAccess;
+import com.graphhopper.routing.profiles.RoadClass;
+import com.graphhopper.routing.profiles.RoadEnvironment;
+import com.graphhopper.routing.profiles.Surface;
+import com.graphhopper.routing.util.parsers.OSMMaxSpeedParser;
+import com.graphhopper.routing.util.parsers.OSMMaxWidthParser;
+import com.graphhopper.routing.util.parsers.OSMRoadAccessParser;
+import com.graphhopper.routing.util.parsers.OSMRoadClassParser;
+import com.graphhopper.routing.util.parsers.OSMRoadEnvironmentParser;
+import com.graphhopper.routing.util.parsers.OSMSurfaceParser;
+import com.graphhopper.routing.util.parsers.SpatialRuleParser;
 import com.graphhopper.routing.util.spatialrules.SpatialRule;
 import com.graphhopper.routing.util.spatialrules.SpatialRuleLookup;
 import com.graphhopper.routing.util.spatialrules.countries.GermanySpatialRule;
@@ -17,13 +42,6 @@ import com.graphhopper.util.TranslationMapTest;
 import com.graphhopper.util.shapes.BBox;
 import com.graphhopper.util.shapes.GHPoint;
 import com.graphhopper.util.shapes.Polygon;
-import org.junit.Test;
-
-import java.util.Arrays;
-import java.util.Collections;
-
-import static com.graphhopper.util.GHUtility.updateDistancesFor;
-import static org.junit.Assert.*;
 
 /**
  * @author Peter Karich
@@ -332,27 +350,6 @@ public class DataFlagEncoderTest {
     public void stringToMeterException() {
         // Unexpected values
         OSMMaxWidthParser.stringToMeter("height limit 1.5m");
-    }
-
-    @Test
-    public void stringToTons() {
-        assertEquals(1.5, OSMWeightExtractor.stringToTons("1.5"), DELTA);
-        assertEquals(1.5, OSMWeightExtractor.stringToTons("1.5 t"), DELTA);
-        assertEquals(1.5, OSMWeightExtractor.stringToTons("1.5   t"), DELTA);
-        assertEquals(1.5, OSMWeightExtractor.stringToTons("1.5 tons"), DELTA);
-        assertEquals(1.5, OSMWeightExtractor.stringToTons("1.5 ton"), DELTA);
-        assertEquals(1.5, OSMWeightExtractor.stringToTons("3306.9 lbs"), DELTA);
-        assertEquals(3, OSMWeightExtractor.stringToTons("3 T"), DELTA);
-        assertEquals(3, OSMWeightExtractor.stringToTons("3ton"), DELTA);
-
-        // maximum gross weight
-        assertEquals(6, OSMWeightExtractor.stringToTons("6t mgw"), DELTA);
-    }
-
-    @Test(expected = NumberFormatException.class)
-    public void stringToTonsException() {
-        // Unexpected values
-        OSMWeightExtractor.stringToTons("weight limit 1.5t");
     }
 
     @Test
